requirements.txt
appnope==0.1.4
astropy==6.1.7
astropy-iers-data==0.2025.6.16.0.38.47
asttokens==3.0.0
certifi==2025.6.15
charset-normalizer==3.4.2
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
exceptiongroup==1.3.0
executing==2.2.0
filelock==3.18.0
fonttools==4.58.4
fsspec==2025.5.1
hf-xet==1.1.4
huggingface-hub==0.33.0
idna==3.10
ipykernel==6.29.5
ipython==8.37.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==1.26.4
packaging==25.0
pandas==2.3.0
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
prompt_toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyerfa==2.0.1.5
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.0
requests==2.32.4
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
six==1.17.0
stack-data==0.6.3
sympy==1.14.0
threadpoolctl==3.6.0
timm==1.0.15
torch==2.2.2
torchvision==0.17.2
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
wcwidth==0.2.13
